chore: Upgrade for Flarum 2.0
