+++ composer.json
@@ -19,7 +19,7 @@
         }
     ],
     "require": {
-        "flarum/core": "^2.0.0-beta",
+        "flarum/core": "^2.0.0",
         "guzzlehttp/guzzle": "^7.1"
     },
     "authors": [
@@ -28,6 +28,10 @@
             "email": "me+fof@datitisev.me",
             "homepage": "https://datitisev.me",
             "role": "Developer"
+        },
+        {
+            "name": "IanM",
+            "email": "ian@flarum.org"
         }
     ],
     "autoload": {
@@ -61,8 +65,8 @@
         }
     },
     "require-dev": {
-        "flarum/phpstan": "^2.0.0-beta",
-        "flarum/testing": "^2.0.0-beta"
+        "flarum/phpstan": "^2.0.0",
+        "flarum/testing": "^2.0.0"
     },
     "scripts": {
         "analyse:phpstan": "phpstan analyse",
