+++ phpstan.neon
@@ -2,8 +2,7 @@ includes:
   - vendor/flarum/phpstan/extension.neon
 
 parameters:
-  # The level will be increased in Flarum 2.0
-  level: 5
+  level: 6
   paths:
     - extend.php
     - src
